Revert "PR: ItemScoreBoard 기능 추가, 화면 해상도 변경"
