--- src/main/java/save/ItemScoreList.json
@@ -1,1 +1,0 @@
-{"score0":20,"score1":20,"name1":"50","mode1":"easy","name0":"50","mode0":"easy"}
