+++ src/main/java/save/ScreenSize.json
@@ -1,1 +1,1 @@
-{"width":800,"height":600}
+{"width":400,"height":500}
