+++ src/main/java/setting/KeySetting.java
@@ -21,10 +21,6 @@ public class KeySetting {
 
     private int oneTimeDown;
 
-    private int escape;
-
-    private int keyNum=7;
-
     public KeySetting() {
         this.left = 37;
         this.right = 39;
@@ -32,8 +28,6 @@ public class KeySetting {
         this.downBlock = 40;
         this.stop = 84;
         this.oneTimeDown = 32;
-        this.escape=27;
-
     }
 
     public void resetDefault(){
@@ -43,37 +37,6 @@ public class KeySetting {
         this.downBlock = 40;
         this.stop = 84;
         this.oneTimeDown = 32;
-        this.escape=27;
-    }
-
-    //key 중복값 있는지 확인
-    public boolean overLapKeySetting(){
-        boolean flag=false;
-
-        int[] checkOverlap = new int[keyNum];
-
-        checkOverlap[0] = this.left;
-        checkOverlap[1]=this.right;
-        checkOverlap[2] = this.turnBlock;
-        checkOverlap[3] = this.downBlock;
-        checkOverlap[4] = this.oneTimeDown;
-        checkOverlap[5] = this.stop;
-        checkOverlap[6] = this.escape;
-
-        for (int i = 0; i < checkOverlap.length; i++) {
-            for (int j = i + 1; j < checkOverlap.length - 1; j++) {
-                if (checkOverlap[i] == checkOverlap[j]) {
-                    flag=true;
-                }
-            }
-            if (flag) {
-                break;
-            }
-        }
-
-        return flag;
-
-
     }
 
     public static KeySetting getKeySetting() {
@@ -104,46 +67,20 @@ public class KeySetting {
         return oneTimeDown;
     }
 
-    public void setKeySetting(int left, int right, int turnBlock, int downBlock, int stop, int oneTimeDown,int escape) {
+    public void setKeySetting(int left, int right, int turnBlock, int downBlock, int stop, int oneTimeDown) {
         this.left = left;
         this.right = right;
         this.turnBlock = turnBlock;
         this.downBlock = downBlock;
         this.stop = stop;
         this.oneTimeDown = oneTimeDown;
-        this.escape = escape;
     }
 
 
-    public void setLeft(int left) {
-        this.left = left;
-    }
 
-    public void setRight(int right) {
-        this.right = right;
-    }
 
-    public void setTurnBlock(int turnBlock) {
-        this.turnBlock = turnBlock;
-    }
 
-    public void setDownBlock(int downBlock) {
-        this.downBlock = downBlock;
-    }
 
-    public void setStop(int stop) {
-        this.stop = stop;
-    }
 
-    public void setOneTimeDown(int oneTimeDown) {
-        this.oneTimeDown = oneTimeDown;
-    }
 
-    public int getEscape() {
-        return escape;
-    }
-
-    public void setEscape(int escape) {
-        this.escape = escape;
-    }
 }
