+++ src/main/java/setting/SaveAndLoad.java
@@ -3,9 +3,9 @@ package setting;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.json.simple.JSONObject;
 import org.json.simple.parser.JSONParser;
-import scoreBoard.NoItemScoreBoard.ScoreList;
+import org.json.simple.parser.ParseException;
+import scoreBoard.ScoreList;
 import scoreBoard.User;
-import scoreBoard.scoreBoradItem.ItemScoreList;
 
 import java.io.*;
 import java.util.HashMap;
@@ -14,7 +14,6 @@ import java.util.List;
 public class SaveAndLoad {
     private static ScreenSize screenSize = ScreenSize.getInstance();
     private static ScoreList scoreList = ScoreList.getInstance();
-    private static ItemScoreList itemScoreList = ItemScoreList.getInstance();
     private static KeySetting keySetting = KeySetting.getInstance();
 
 
@@ -33,9 +32,8 @@ public class SaveAndLoad {
         keySettingJson.put("turnBlock", keySetting.getTurnBlock());
         keySettingJson.put("oneTimeDown", keySetting.getOneTimeDown());
         keySettingJson.put("stop", keySetting.getStop());
-        keySettingJson.put("escape", keySetting.getEscape());
 
-        //스코어보드(item모드 없음)
+        //스코어보드
         List<User> list = scoreList.getList();
         JSONObject scoreBoardJson = new JSONObject();
         int size = list.size();
@@ -46,22 +44,6 @@ public class SaveAndLoad {
         for (int i=0;i<size;i++){
             scoreBoardJson.put("name"+i, list.get(i).getName());
             scoreBoardJson.put("score"+i, list.get(i).getScore());
-            scoreBoardJson.put("mode"+i, list.get(i).getMode());
-        }
-
-        //스코어보드(item모드)
-        List<User> listForItem = itemScoreList.getList();
-        JSONObject scoreBoardItemJson = new JSONObject();
-        System.out.println("itemScoreList = " + itemScoreList);
-        int sizeForItemList = listForItem.size();
-        //10개 까지만 저장
-        if(sizeForItemList>=10){
-            sizeForItemList=10;
-        }
-        for (int i=0;i<sizeForItemList;i++){
-            scoreBoardItemJson.put("name"+i, listForItem.get(i).getName());
-            scoreBoardItemJson.put("score"+i, listForItem.get(i).getScore());
-            scoreBoardItemJson.put("mode"+i, listForItem.get(i).getMode());
         }
 
         try{
@@ -78,17 +60,10 @@ public class SaveAndLoad {
             keySettingFile.close();
 
             //스코어리스트 10개까지 저장
-            FileWriter scoreListFile = new FileWriter("src/main/java/save/NoItemScoreList.json");
+            FileWriter scoreListFile = new FileWriter("src/main/java/save/ScoreList.json");
             scoreListFile.write(scoreBoardJson.toJSONString());
             scoreListFile.flush();
             scoreListFile.close();
-
-            //Item스코어리스트 10개까지 저장
-            FileWriter ItemScoreListFile = new FileWriter("src/main/java/save/ItemScoreList.json");
-            ItemScoreListFile.write(scoreBoardItemJson.toJSONString());
-            ItemScoreListFile.flush();
-            ItemScoreListFile.close();
-
 
 
         } catch (IOException e) {
@@ -101,38 +76,31 @@ public class SaveAndLoad {
 
         try{
             //화면 크기 load
+//            FileReader readerToScreenSize = new FileReader("src/main/java/save/ScreenSize.json");
+//            Object screenSizeJson1 = parser.parse(readerToScreenSize);
+//            JSONObject screenSizeJson2 =(JSONObject) screenSizeJson1;
+//            readerToScreenSize.close();
             InputStream getScreenSize = new FileInputStream("src/main/java/save/ScreenSize.json");
             HashMap<String,Object> screenSizeMap = new ObjectMapper().readValue(getScreenSize, HashMap.class);
             screenSize.setWidth((Integer) screenSizeMap.get("width"));
             screenSize.setHeight((Integer) screenSizeMap.get("height"));
 
             //스코어 보드 load
-            InputStream getScoreBoard = new FileInputStream("src/main/java/save/NoItemScoreList.json");
+//            FileReader readerToScoreList = new FileReader("src/main/java/save/ScoreList.json");
+//            Object scoreListJson1 = parser.parse(readerToScoreList);
+//            JSONObject scoreListJson2 =(JSONObject) scoreListJson1;
+//            readerToScoreList.close();
+            InputStream getScoreBoard = new FileInputStream("src/main/java/save/ScoreList.json");
             HashMap<String,Object> scoreBoardMap = new ObjectMapper().readValue(getScoreBoard, HashMap.class);
             System.out.println("scoreBoardMap = " + scoreBoardMap.size());
-            int size = scoreBoardMap.size()/3;
+            int size = scoreBoardMap.size()/2;
             if (size >= 10) {
                 size=10;
             }
             for (int i = 0; i < size; i++) {
-                scoreList.push(new User((String) scoreBoardMap.get("name" + i), (Integer) scoreBoardMap.get("score" + i), (String)scoreBoardMap.get("mode" + i)));
-
+                scoreList.push(new User((String) scoreBoardMap.get("name" + i), (Integer) scoreBoardMap.get("score" + i)));
             }
             scoreList.sortDescByScore();
-
-            //Item스코어 보드 load
-            InputStream getItemScoreBoard = new FileInputStream("src/main/java/save/ItemScoreList.json");
-            HashMap<String,Object> ItemscoreBoardMap = new ObjectMapper().readValue(getItemScoreBoard, HashMap.class);
-            System.out.println("scoreBoardMap = " + ItemscoreBoardMap.size());
-            int ItemListSize = ItemscoreBoardMap.size()/3;
-            if (ItemListSize >= 10) {
-                ItemListSize=10;
-            }
-            for (int i = 0; i < ItemListSize; i++) {
-                itemScoreList.push(new User((String) ItemscoreBoardMap.get("name" + i), (Integer) ItemscoreBoardMap.get("score" + i), (String)ItemscoreBoardMap.get("mode" + i)));
-
-            }
-            itemScoreList.sortDescByScore();
 
             //keySetting Load
             InputStream getKeySetting = new FileInputStream("src/main/java/save/KeySetting.json");
@@ -143,8 +111,7 @@ public class SaveAndLoad {
             int turnBlock = (Integer)getKeySettingMap.get("turnBlock");
             int stop = (Integer)getKeySettingMap.get("stop");
             int oneTimeDown = (Integer)getKeySettingMap.get("oneTimeDown");
-            int escape = (Integer)getKeySettingMap.get("escape");
-            keySetting.setKeySetting(left,right,turnBlock,downBlock,stop,oneTimeDown,escape);
+            keySetting.setKeySetting(left,right,turnBlock,downBlock,stop,oneTimeDown);
