+++ test/scoreBoard/ScoreListTest.java
@@ -1,10 +1,11 @@
 
 package scoreBoard;
 
+import junit.framework.Assert; //junit framework
+import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import scoreBoard.NoItemScoreBoard.ScoreList;
 
 import static org.assertj.core.api.Assertions.*;
 
@@ -20,8 +21,8 @@ public class ScoreListTest {
     @DisplayName("점수 모두 삭제")
     public void 점수모두삭제(){
         //given
-        User test1 = new User("test1", 100,"easy");
-        User test2 = new User("test2", 100,"easy");
+        User test1 = new User("test1", 100);
+        User test2 = new User("test2", 100);
         //when
         scoreList.deleteAll();
         //then
@@ -33,8 +34,8 @@ public class ScoreListTest {
     @DisplayName("점수  push  기능")
     public void 점수푸쉬(){
         //given
-        User test1 = new User("test1", 100,"easy");
-        User test2 = new User("test2", 100,"easy");
+        User test1 = new User("test1", 100);
+        User test2 = new User("test2", 100);
         //when
         scoreList.push(test1);
         scoreList.push(test2);
@@ -48,9 +49,9 @@ public class ScoreListTest {
     @DisplayName("점수 오름차순 정렬")
     public void ScoreListSortByDescTest() {
         //given
-        User user1 = new User("test1", 100,"easy");
-        User user2 = new User("test2", 60,"easy");
-        User user3 = new User("test3", 80,"easy");
+        User user1 = new User("test1", 100);
+        User user2 = new User("test2", 60);
+        User user3 = new User("test3", 80);
         scoreList.push(user1);
         scoreList.push(user2);
         scoreList.push(user3);
